Move cli-specific code to another file
